Refactor to incorporate PrisonLocator into Prison class
I have refactored the Prison class to incorporate a `.find` class method, removing the need for the PrisonLocator class.

I've also relocated the prison data into a `prisons.yaml` file which is loaded in at runtime, rather than being hardcoded into the Prison class itself. This will improve maintainability of the prison data, and will make it easier to abstract to an external data source if necessary in the future.