Initial test for checking a vacancies' validity
ie. Is it at a prison that we know.
There can be more than one prison listed in a vacancy, so we
*have to* check the entire string rather than merely checking that the
text contains a valid prison.

This may yield false negatives (by design) -  anything that is not boiler plate text
NOR a valid prison must be surfaced as 'bad data', since it *may be* a misspelled prison.